Move scan method into library
