+++ src/main.rs
@@ -1,8 +1,7 @@
 use std::io::{Error, Write};
 use std::{env, io,
-          net::{IpAddr, SocketAddr, ToSocketAddrs},
-          time::Duration,};
-use futures::{stream, StreamExt};
+          net::{SocketAddr, ToSocketAddrs},
+};
 use ansi_term::Colour;
 use clap::{load_yaml, App, ArgMatches};
 use dprint_core::formatting::PrintOptions;
@@ -71,7 +70,7 @@ async fn main() -> Result<(), Error> {
                                     .unwrap()
                                     .parse::<u64>()
                                     .unwrap_or(1);
-                scan(socket_addresses[0].ip(), 1000, timeout).await;
+                libfrizz::scan(socket_addresses[0].ip(), 1000, timeout).await;
                 return Ok(());
             }
 
@@ -169,26 +168,5 @@ async fn execute_http(
         out_writer
             .write(format!("{}", Colour::White.paint(body)).as_bytes())
             .ok();
-    }
-}
-
-
-async fn scan(target: IpAddr, concurrency: usize, timeout: u64) {
-    let ports = stream::iter(1..=u16::MAX);
-
-    ports
-        .for_each_concurrent(concurrency, |port| scan_port(target, port, timeout))
-        .await;
-}
-
-async fn scan_port(target: IpAddr, port: u16, timeout: u64) {
-    let timeout = Duration::from_secs(timeout);
-    let socket_address = SocketAddr::new(target, port);
-
-    if tokio::time::timeout(timeout, TcpStream::connect(&socket_address))
-        .await
-        .is_ok()
-    {
-        println!("{}", port);
     }
 }
